% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renviron-package.R
\docType{package}
\name{renviron-package}
\alias{renviron}
\alias{renviron-package}
\title{Administración de archivos de entorno para R}
\description{
Lectura aislada, carga selectiva y edición por clave de archivos de entorno.
Consulte \code{\link[=renviron_read]{renviron_read()}} para lectura y \code{\link[=renviron_add]{renviron_add()}} para edición.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adatar-do.github.io/renviron/}
  \item \url{https://github.com/adatar-do/renviron}
  \item Report bugs at \url{https://github.com/adatar-do/renviron/issues}
}

}
\author{
\strong{Maintainer}: Daniel E. de la Rosa \email{me@dnldelarosa.com} (\href{https://orcid.org/0000-0002-5965-5277}{ORCID})

Other contributors:
\itemize{
  \item Adatar [copyright holder]
}

}
\keyword{internal}
