% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renviron_exists.R
\name{renviron_exists}
\alias{renviron_exists}
\title{Check if an Environment Variable Exists}
\usage{
renviron_exists(key, .renviron = NULL, scope = c("project", "user"))
}
\arguments{
\item{key}{A character string specifying the name of the environment variable to check.}

\item{.renviron}{An optional named list of environment variables to check.
If not provided, the function will load the variables from the .Renviron file using \code{renviron_load()}.}

\item{scope}{A character vector specifying the scope(s) to search for the .Renviron file when loading
environment variables. This is used only when \code{.renviron} is \code{NULL}. Valid values
are "user" and "project". The function searches in the order provided. The default order
is \code{c("project", "user")}.}
}
\value{
TRUE if the environment variable exists, FALSE otherwise.
}
\description{
This function checks for the existence of a specified environment variable in the .Renviron file
or a provided named list of environment variables.
}
\examples{
\dontrun{
# Check if the CENSUS_API_KEY variable exists in the .Renviron file
exists <- renviron_exists("CENSUS_API_KEY")
print(exists)

# Check for the variable in a provided list
env_list <- list(CENSUS_API_KEY='12345', GITHUB_PAT='abcde')
exists_in_list <- renviron_exists("CENSUS_API_KEY", .renviron = env_list)
print(exists_in_list)
}

}
