% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renviron_get.R
\name{renviron_get}
\alias{renviron_get}
\title{Retrieve a Specific Environment Variable from the .Renviron File}
\usage{
renviron_get(key, .renviron = NULL)
}
\arguments{
\item{key}{A character string specifying the name of the environment variable to retrieve.}

\item{.renviron}{An optional named list of environment variables. If provided, the function will
use this list instead of loading the .Renviron file via \code{renviron_load()}. This can be useful
for testing or when working with a modified set of environment variables that has not been saved back to the file.}
}
\value{
The value of the environment variable corresponding to \code{key}, if it exists.
Returns \code{NULL} if the key does not exist in the list.
}
\description{
This function fetches the value of a specified environment variable from the .Renviron file.
It relies on \code{renviron_load()} to load the environment variables into a named list, from which
the value of the specified key is retrieved. If the \code{.renviron} argument is provided, it uses
this list directly, bypassing the need to load from the file.
}
\examples{
\dontrun{
# Assuming the .Renviron file contains:
# CENSUS_API_KEY='12345'

# Retrieve the CENSUS_API_KEY value
api_key <- renviron_get("CENSUS_API_KEY")
print(api_key)
}

}
