% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renviron_list.R
\name{renviron_list}
\alias{renviron_list}
\title{List and mask environment variables}
\usage{
renviron_list(.renviron = NULL, ...)
}
\arguments{
\item{.renviron}{An optional named list of environment variables that should be listed. If not
provided, the function will load the variables from the .Renviron file using \code{renviron_load()}
and mask their values according to the specified scope.}

\item{...}{Additional arguments that are passed to the \code{renviron_load()} function:
- \code{scope}: A character vector specifying the scope(s) to search for the .Renviron file.
Valid options are "user" and "project", with "project" typically having precedence unless
otherwise specified. This determines which .Renviron file the variables are loaded from
if \code{.renviron} is not provided. The default is \code{c("project", "user")}.
- \code{.file}: Specifies the filename to be considered as the environment file within the
specified scope. Default is ".Renviron".}
}
\value{
A named list of all environment variables with their values masked as "*****".
This list includes all variables found in the specified scope if \code{.renviron} is not provided,
or from the provided list otherwise. This method ensures that no sensitive data is displayed,
while still allowing users to understand which variables are set.
}
\description{
This function provides a secure way to view all environment variables from the .Renviron file
or a provided list, with their values masked for security purposes. It can operate over both
user and project scopes as defined by the \code{scope} argument when loading from the .Renviron file.
This feature is useful for debugging or auditing without compromising sensitive information by
exposing actual values of the variables.
}
\examples{
\dontrun{
# List and mask all environment variables from the .Renviron file within the default scope
renviron_list()

# Directly list and mask variables from a custom provided named list
custom_env <- list(API_KEY = "12345", SECRET = "s3cr3t")
masked_env <- renviron_list(.renviron = custom_env)
print(masked_env)
}

}
