% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{renviron_load}
\alias{renviron_load}
\title{Load Environment Variables from the .Renviron File}
\usage{
renviron_load()
}
\value{
A named list where each element corresponds to an environment variable loaded from
the .Renviron file. The list's names are the environment variable names, and the
values are the corresponding variable values, with outer quotes removed if present.
}
\description{
This function reads the .Renviron file determined by \code{renviron_path()}, which considers
both user and project scopes. It converts the variables defined within the file into a
named list in R. Variables should be in the \code{KEY=VALUE} format, where \code{KEY} becomes
the list name, and \code{VALUE} is the corresponding list value. Outer quotes around the value
are optionally removed.
}
\examples{
\dontrun{
# Assuming the .Renviron file contains:
# CENSUS_API_KEY='12345'
# GITHUB_PAT='abcde'

# Load environment variables from the .Renviron file
environment <- renviron_load()

# Access a specific variable
print(environment$CENSUS_API_KEY)
}

}
