% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{renviron_path}
\alias{renviron_path}
\title{Retrieve the Path to the .Renviron File}
\usage{
renviron_path()
}
\value{
A character string representing the path to the .Renviron file that should be loaded.
If no file is found in either the project or user scope, the function may return \code{NULL}.
}
\description{
This function utilizes \code{usethis:::scoped_path_r} to determine the path to the .Renviron file
by considering both user and project scopes. The function prioritizes the project-specific
.Renviron file if present; otherwise, it falls back to the user-level .Renviron.
}
\details{
The search order and precedence are determined by \code{usethis:::scoped_path_r}, which checks
the project directory followed by the user's home directory for the .Renviron file.
}
\examples{
\dontrun{
# To get the path to the currently active .Renviron file:
env_path <- renviron_path()
print(env_path)
}

}
