% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renviron_unset_all.R
\name{renviron_unset_all}
\alias{renviron_unset_all}
\title{Unset all environment variables from the environment file}
\usage{
renviron_unset_all(...)
}
\arguments{
\item{...}{Additional arguments to \code{renviron_load()}:
- \code{scope}: A character vector specifying the scope(s) to search for the .Renviron file.
Valid values are "user" and "project". The function searches in the order provided.
The default order is \code{c("project", "user")}.
- \code{.file}: Optionally specify a different filename to use within the specified scope.
Defaults to ".Renviron".}
}
\value{
Invisibly returns NULL after unsetting the variables.
}
\description{
This function unsets all environment variables that are loaded from the specified
.Renviron file within the given scope. It is particularly useful for cleaning up
the environment in R sessions during testing or after loading configurations that
are no longer needed.
}
\examples{
\dontrun{
renviron_unset_all(scope = "project")
renviron_unset_all(scope = "user", .file = "custom.env")
}

}
